Conflicts:
	.gitignore
	configure.ac
	interfaces/R/Makefile
	interfaces/functions.def
	src/Makefile.am

+++ src/igraph_fixed_vectorlist.c
@@ -52,17 +52,17 @@ int igraph_fixed_vectorlist_convert(igraph_fixed_vectorlist_t *l,
   IGRAPH_VECTOR_INIT_FINALLY(&sizes, size);
 
   for (i=0; i<no; i++) {
-    long int to=VECTOR(*from)[i];
+    long int to=(long int) VECTOR(*from)[i];
     if (to >= 0) { VECTOR(sizes)[to] += 1; }
   }
   for (i=0; i<size; i++) {
     igraph_vector_t *v=&(l->vecs[i]);
-    IGRAPH_CHECK(igraph_vector_init(v, VECTOR(sizes)[i]));
+    IGRAPH_CHECK(igraph_vector_init(v, (long int) VECTOR(sizes)[i]));
     igraph_vector_clear(v);
     VECTOR(l->v)[i]=v;
   }
   for (i=0; i<no; i++) {
-    long int to=VECTOR(*from)[i];
+    long int to=(long int) VECTOR(*from)[i];
     if (to >= 0) { 
       igraph_vector_t *v=&(l->vecs[to]);
       igraph_vector_push_back(v, i);
